+++ spec/lightstreamer/session_spec.rb
@@ -125,19 +125,19 @@ describe Lightstreamer::Session do
       expect(Lightstreamer::PostRequest).to receive(:request_body)
         .with(LS_session: 'session', LS_op: :add, LS_table: subscription.id, LS_mode: 'MERGE', LS_id: %w(item),
               LS_schema: %w(field), LS_selector: 'selector', LS_data_adapter: nil, LS_requested_max_frequency: 0.0,
-              LS_snapshot: false)
+              LS_snapshot: true)
         .and_return('body1')
 
       expect(Lightstreamer::PostRequest).to receive(:request_body)
         .with(LS_session: 'session', LS_op: :add, LS_table: second_subscription.id, LS_mode: 'MERGE', LS_schema: [],
-              LS_id: [], LS_selector: nil, LS_data_adapter: nil, LS_requested_max_frequency: 0.0, LS_snapshot: false)
+              LS_id: [], LS_selector: nil, LS_data_adapter: nil, LS_requested_max_frequency: 0.0, LS_snapshot: true)
         .and_return('body2')
 
       expect(Lightstreamer::PostRequest).to receive(:bulk_execute)
         .with('http://a.com/lightstreamer/control.txt', %w(body1 body2))
         .and_return([nil, Lightstreamer::Errors::InvalidDataAdapterError.new])
 
-      errors = session.bulk_subscription_start subscription, second_subscription
+      errors = session.bulk_subscription_start [subscription, second_subscription], snapshot: true
 
       expect(errors.size).to eq(2)
       expect(errors[0]).to be nil